requirements.txt
asgiref==3.3.1
Django==3.1.6
django-cors-headers==3.7.0
django-restframework==0.0.1
pytz==2021.1
sqlparse==0.4.1
